+++ src/utils/temporal/index.ts
@@ -1,37 +1,31 @@
 type DateValue = string | number | Date;
-export function current(d?: DateValue): Date {
+export function current(d?: DateValue) {
 	if (d instanceof Date) return d;
 	return d ? new Date(d) : new Date();
 }
 
-const word = {
-	days: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
-	months: [
-		'January',
-		'February',
-		'March',
-		'April',
-		'May',
-		'June',
-		'July',
-		'August',
-		'September',
-		'October',
-		'November',
-		'December',
-	],
-};
-
 const str = (v: DateValue) => `${v}`;
 const pad = (v: DateValue, len = 2) => str(v).padStart(len, '0');
 
+export const i18n: Record<string, { days: string[]; months: string[] }> = {
+	en: {
+		days: 'Sunday|Monday|Tuesday|Wednesday|Thursday|Friday|Saturday'.split('|'),
+		months: [
+			...'January|February|March|April|May|June|July'.split('|'),
+			...'August|September|October|November|December'.split('|'),
+		],
+	},
+};
+
 interface BuildOptions {
 	base?: 'UTC';
+	lang?: string;
 }
-export function build({ base }: BuildOptions) {
+export function build({ base, lang }: BuildOptions) {
 	const method = base === 'UTC' ? 'getUTC' : 'get';
+	const { days, months } = i18n[lang || 'en'];
 
-	return (date?: DateValue) => {
+	return function formatter(date?: DateValue) {
 		const d = current(date);
 		if (Number.isNaN(+d)) throw SyntaxError('Invalid Date');
 
@@ -54,13 +48,13 @@ export function build({ base }: BuildOptions) {
 		const tokens = {
 			D: () => str(now.date()),
 			DD: () => pad(now.date()),
-			DDD: () => word.days[now.day()].slice(0, 3),
-			DDDD: () => word.days[now.day()],
+			DDD: () => days[now.day()].slice(0, 3),
+			DDDD: () => days[now.day()],
 
 			M: () => str(now.month() + 1),
 			MM: () => pad(now.month() + 1),
-			MMM: () => word.months[now.month()].slice(0, 3),
-			MMMM: () => word.months[now.month()],
+			MMM: () => months[now.month()].slice(0, 3),
+			MMMM: () => months[now.month()],
 
 			YY: () => str(now.year()).slice(2),
 			YYYY: () => str(now.year()),
@@ -83,7 +77,7 @@ export function build({ base }: BuildOptions) {
 			ZZZ: () => `${sign}${pad(tz[0])}:${tz[1]}`,
 		};
 
-		return (mask = 'DDDD, DD MMMM YYYY') => {
+		return function renderer(mask = 'DDDD, DD MMMM YYYY') {
 			const EXP = /D{1,4}|M{1,4}|YY(?:YY)?|([hHmsAPap])\1?|Z{1,3}|\[([^\]\[]|\[[^\[\]]*\])*\]/g;
 			return mask.replace(EXP, ($) => {
 				const exe = tokens[$ as keyof typeof tokens];