+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # mauss changelog
 
+## 0.4.12 - 2023/05/02
+
+- ([#224](https://github.com/alchemauss/mauss/pull/224)) add string guards
+- ([#221](https://github.com/alchemauss/mauss/pull/221)) add `tsf` function to `/std` module
+
 ## 0.4.11 - 2023/03/30
 
 - ([#216](https://github.com/alchemauss/mauss/pull/216)) remove deprecated option for TS 5.0